adding crud of calendar backend in frontend

+++ src/hooks/useCalendarStore.js
@@ -1,32 +1,61 @@
 import { useDispatch, useSelector } from 'react-redux';
-import { onAddNewEvent, onDeleteEvent, onSetActiveEvent, onUpdateEvent } from '../store/calendar/calendarSlice';
-
+import { onAddNewEvent, onDeleteEvent, onLoadEvent, onSetActiveEvent, onUpdateEvent } from '../store/calendar/calendarSlice';
+import calendarApi from '../api/calendarApi';
+import { convertEventsToDateEvents } from '../helpers';
+import Swal from "sweetalert2";
+import 'sweetalert2/dist/sweetalert2.min.css'
 
 export const useCalendarStore = () => {
   
     const dispatch = useDispatch();
     const { events, activeEvent } = useSelector( state => state.calendar );
+    const { user } = useSelector( state => state.auth );
 
     const setActiveEvent = ( calendarEvent ) => {
         dispatch( onSetActiveEvent( calendarEvent ) )
     }
 
     const startSavingEvent = async( calendarEvent ) => {
-        // TODO: llegar al backend
-
-        // Todo bien
-        if( calendarEvent._id ) {
-            // Actualizando
-            dispatch( onUpdateEvent({ ...calendarEvent }) );
-        } else {
+        try {
+            if( calendarEvent.id ) {
+                // Actualizando
+                await calendarApi.put(`/calendar/updateEvent/${calendarEvent.id}`, calendarEvent)
+                dispatch( onUpdateEvent({ ...calendarEvent, user }) );
+                return;
+            }
             // Creando
-            dispatch( onAddNewEvent({ ...calendarEvent, _id: new Date().getTime() }) );
+            const { data } = await calendarApi.post('/calendar/addEvent', calendarEvent)
+            console.log({data});
+            dispatch( onAddNewEvent({ ...calendarEvent, id: data.event.id, user: user}) );
+        } catch (error) {
+            console.log(error);
+            Swal.fire('Error al guardar', error.data?.message, 'error')      
         }
     }
 
-    const startDeletingEvent = () => {
+    const startDeletingEvent = async() => {
         // Todo: Llegar al backend
-        dispatch( onDeleteEvent() );
+        try {
+            if( activeEvent.id ) {
+                await calendarApi.delete(`/calendar/deleteEvent/${activeEvent.id}`)
+                dispatch( onDeleteEvent() );
+            }
+            
+        } catch (error) {
+            console.log(error);
+            Swal.fire('Error al eliminar', error.data?.message, 'error')      
+        }
+    }
+
+    const startLoadingEvents = async() => {
+        try {
+            const { data } = await calendarApi.get('/calendar/getEvents')
+            const events = convertEventsToDateEvents( data.events )
+            dispatch( onLoadEvent(events) )
+        } catch (error) {
+            console.log('Error cargando eventos');
+            console.log(error);
+        }
     }
 
 
@@ -40,5 +69,6 @@ export const useCalendarStore = () => {
         startDeletingEvent,
         setActiveEvent,
         startSavingEvent,
+        startLoadingEvents
     }
 }
